Necessary Android Studio Set Up for Project
Note: I admit that this set up seems to have an issue ("Android Wear has stopped running"), but if you ignore the first 3-5 errors, the app itself will function. Probably not the best sign and has to do with so many features being disabled (as they wouldn't be necessary for an insulin pump). (Could also just be on my end, let me know if it affects any of you as well.)